separated authentication into it's own file
